+++ src/main/java/dh/meli/projeto_integrador/dto/dtoInput/CartDto.java
@@ -1,4 +1,4 @@
-package dh.meli.projeto_integrador.dtos.dtoInput;
+package dh.meli.projeto_integrador.dto.dtoInput;
 
 import dh.meli.projeto_integrador.enumClass.PurchaseOrderStatusEnum;
 import lombok.AllArgsConstructor;
