+++ src/main/java/dh/meli/projeto_integrador/service/ICartService.java
@@ -1,7 +1,7 @@
 package dh.meli.projeto_integrador.service;
 
-import dh.meli.projeto_integrador.dtos.dtoInput.CartDto;
-import dh.meli.projeto_integrador.dtos.dtoOutput.TotalPriceDto;
+import dh.meli.projeto_integrador.dto.dtoInput.CartDto;
+import dh.meli.projeto_integrador.dto.dtoOutput.TotalPriceDto;
 
 public interface ICartService {
     // método para criação de um carrinho de compras
