+++ src/test/java/dh/meli/projeto_integrador/controller/OrderControllerTest.java
@@ -1,7 +1,7 @@
 package dh.meli.projeto_integrador.controller;
 
-import dh.meli.projeto_integrador.dto.BatchDto;
-import dh.meli.projeto_integrador.dto.OrderEntryDto;
+import dh.meli.projeto_integrador.dto.dtoInput.BatchDto;
+import dh.meli.projeto_integrador.dto.dtoInput.OrderEntryDto;
 import dh.meli.projeto_integrador.service.OrderService;
 import dh.meli.projeto_integrador.util.Generators;
 import org.junit.jupiter.api.BeforeEach;
